requirements.txt
numpy==1.17.2
opencv_python==4.4.0.42
dlib==19.21.0
pandas==0.25.1
Pillow==7.2.0
